+++ test/acceptance/multilayer.js
@@ -80,7 +80,7 @@ suite('multilayer', function() {
         });
     });
 
-    test("layergroup with 2 layers, each with its style, grid on first", function(done) {
+    test("layergroup with 2 layers, each with its style", function(done) {
 
       var layergroup =  {
         version: '1.0.1',
@@ -88,13 +88,14 @@ suite('multilayer', function() {
            { options: {
                sql: 'select cartodb_id, ST_Translate(the_geom, 50, 0) as the_geom from test_table limit 2',
                cartocss: '#layer { marker-fill:red; marker-width:32; marker-allow-overlap:true; }', 
-               cartocss_version: '2.0.1' ,
+               cartocss_version: '2.0.1',
                interactivity: [ 'cartodb_id' ]
              } },
            { options: {
                sql: 'select cartodb_id, ST_Translate(the_geom, -50, 0) as the_geom from test_table limit 2 offset 2',
                cartocss: '#layer { marker-fill:blue; marker-allow-overlap:true; }', 
-               cartocss_version: '2.0.2' 
+               cartocss_version: '2.0.2',
+               interactivity: [ 'cartodb_id' ]
              } }
         ]
       };
@@ -134,13 +135,13 @@ suite('multilayer', function() {
               });
           });
         },
-        function do_get_grid(err)
+        function do_get_grid0(err)
         {
           if ( err ) throw err;
           var next = this;
           assert.response(server, {
               url: '/database/windshaft_test/layergroup/' + expected_token
-                 + '/0/0/0.grid.json',
+                 + '/layer0/0/0/0.grid.json?interactivity=cartodb_id',
               method: 'GET'
           }, {}, function(res) {
               assert.equal(res.statusCode, 200, res.body);
@@ -151,83 +152,13 @@ suite('multilayer', function() {
               });
           });
         },
-        function finish(err) {
-          var errors = [];
-          if ( err ) errors.push(err.message);
-          redis_client.keys("map_style|windshaft_test|~" + expected_token, function(err, matches) {
-              if ( err ) errors.push(err.message);
-              assert.equal(matches.length, 1, "Missing expected token " + expected_token + " from redis");
-              redis_client.del(matches, function(err) {
-                if ( err ) errors.push(err.message);
-                if ( errors.length ) done(new Error(errors));
-                else done(null);
-              });
-          });
-        }
-      );
-    });
-
-    test("layergroup with 2 layers, each with its style, grid on second, version 1.0.0", function(done) {
-
-      var layergroup =  {
-        version: '1.0.0',
-        layers: [
-           { options: {
-               sql: 'select cartodb_id, ST_Translate(the_geom, 50, 0) as the_geom from test_table limit 2',
-               cartocss: '#layer { marker-fill:red; marker-width:32; marker-allow-overlap:true; }', 
-               cartocss_version: '2.0.1' 
-             } },
-           { options: {
-               sql: 'select cartodb_id, ST_Translate(the_geom, -50, 0) as the_geom from test_table limit 2 offset 2',
-               cartocss: '#layer { marker-fill:blue; marker-allow-overlap:true; }', 
-               cartocss_version: '2.0.2',
-               interactivity: 'cartodb_id'
-             } }
-        ]
-      };
-
-      var expected_token = "a7556dc21542ae8bcf16906e73ad5029";
-      Step(
-        function do_post()
-        {
-          var next = this;
-          assert.response(server, {
-              url: '/database/windshaft_test/layergroup',
-              method: 'POST',
-              headers: {'Content-Type': 'application/json' },
-              data: JSON.stringify(layergroup)
-          }, {}, function(res) {
-              assert.equal(res.statusCode, 200, res.body);
-              var parsedBody = JSON.parse(res.body);
-              if ( expected_token ) assert.deepEqual(parsedBody, {layergroupid: expected_token, layercount: 2});
-              else expected_token = parsedBody.layergroupid;
-              next(null, res);
-          });
-        },
-        function do_get_tile(err)
-        {
-          if ( err ) throw err;
-          var next = this;
-          assert.response(server, {
-              url: '/database/windshaft_test/layergroup/' + expected_token + '/0/0/0.png',
-              method: 'GET',
-              encoding: 'binary'
-          }, {}, function(res) {
-              assert.equal(res.statusCode, 200, res.body);
-              assert.equal(res.headers['content-type'], "image/png");
-              assert.imageEqualsFile(res.body, './test/fixtures/test_table_0_0_0_multilayer1.png', 2,
-                function(err, similarity) {
-                  next(err);
-              });
-          });
-        },
-        function do_get_grid(err)
+        function do_get_grid1(err)
         {
           if ( err ) throw err;
           var next = this;
           assert.response(server, {
               url: '/database/windshaft_test/layergroup/' + expected_token
-                 + '/0/0/0.grid.json',
+                 + '/layer1/0/0/0.grid.json?interactivity=cartodb_id',
               method: 'GET'
           }, {}, function(res) {
               assert.equal(res.statusCode, 200, res.body);
@@ -387,7 +318,7 @@ suite('multilayer', function() {
           var next = this;
           assert.response(server, {
               url: '/database/windshaft_test/layergroup/' + token1
-                 + '/0/0/0.grid.json',
+                 + '/layer0/0/0/0.grid.json?interactivity=cartodb_id',
               method: 'GET'
           }, {}, function(res) {
               assert.equal(res.statusCode, 200, res.body);
@@ -421,7 +352,7 @@ suite('multilayer', function() {
           var next = this;
           assert.response(server, {
               url: '/database/windshaft_test/layergroup/' + token2
-                 + '/0/0/0.grid.json',
+                 + '/layer0/0/0/0.grid.json?interactivity=cartodb_id',
               method: 'GET'
           }, {}, function(res) {
               assert.equal(res.statusCode, 200, res.body);